Move the Xplat to the standard structure
