--- Extensions/XplatGenerateReleaseNotes/V3/tslint.json
@@ -1,60 +1,0 @@
-{
-    "rules": {
-        "class-name": true,
-        "comment-format": [
-            true,
-            "check-space"
-        ],
-        "curly": true,
-        "indent": [
-            true,
-            "spaces"
-        ],
-        "no-consecutive-blank-lines": true,
-        "no-console": false,
-        "no-duplicate-variable": true,
-        "no-eval": false, // needed for this code base usually true
-        "no-internal-module": true,
-        "no-trailing-whitespace": true,
-        "no-var-keyword": false, // should swapout as refactor
-        "one-line": [
-            true,
-            "check-open-brace",
-            "check-whitespace"
-        ],
-        "quotemark": [
-            true,
-            "double"
-        ],
-        "semicolon": [
-            true,
-            "always"
-        ],
-        "triple-equals": [
-            true,
-            "allow-null-check"
-        ],
-        "typedef-whitespace": [
-            true,
-            {
-                "call-signature": "nospace",
-                "index-signature": "nospace",
-                "parameter": "nospace",
-                "property-declaration": "nospace",
-                "variable-declaration": "nospace"
-            }
-        ],
-        "variable-name": [
-            true,
-            "ban-keywords"
-        ],
-        "whitespace": [
-            true,
-            "check-branch",
-            "check-decl",
-            "check-operator",
-            "check-separator",
-            "check-type"
-        ]
-    }
-}
