+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/package.json
@@ -1,11 +1,15 @@
 {
   "name": "XPlatGenerateReleaseNotesTask",
   "scripts": {
-    "build": "npm run lint && npm run transpile ",
-    "lint": "tslint -c tslint.json *.ts test/*.ts testconsole/*.ts",
+    "build": "npm run clean && npm run lint && npm run transpile ",
+    "lint": "tslint -c tslint.json src/*.ts test/*.ts testconsole/*.ts",
+    "package": "npm prune -production && npm run copy",
+    "copy": "ncp ./node_modules ./task/node_modules && ncp ./dist/src ./task",
+    "clean": "rimraf ./dist/src && rimraf ./dist/test && rimraf ./task/*.js && rimraf ./task/*.js.map && rimraf ./task/node_modules",
     "transpile": "tsc -p ./",
-    "test": "mocha -r ts-node/register ./test/*.ts --reporter mocha-junit-reporter --reporter-options mochaFile=./test-output/test-results.xml ",
-    "test-no-logger": "mocha -r ts-node/register ./test/*.ts "
+    "test": "mocha -r ts-node/register ./test/*.test.ts --reporter mocha-junit-reporter --reporter-options mochaFile=./test-output/test-results.xml ",
+    "test-no-logger": "mocha -r ts-node/register ./test/*.test.ts "
+    
   },
   "description": "XPlatGenerateReleaseNotesTask",
   "main": "GenerateReleaseNotes.js",
