+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/src/Issue349Workaround.ts
