+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/src/agentSpecific.ts
