+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/task/icon.png
