+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/task/task.json
