+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/testconsole/build-settings.json
