+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/testconsole/release-settings.json
