+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/tsconfig.json
@@ -4,6 +4,7 @@
        "target": "es6",
        "noImplicitAny": false,
        "moduleResolution": "node",
+       "outDir": "dist",
        "sourceMap":  true,
        "rootDirs": [
           "src/",
