+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/tslint.json
