+++ Extensions/XplatGenerateReleaseNotes/XplatGenerateReleaseNotesTask/typings.json
