--- Extensions/XplatGenerateReleaseNotes/make-util.js
@@ -1,162 +1,0 @@
-var fs = require('fs');
-var check = require('validator');
-var path = require('path');
-var minimatch = require('minimatch');
-var shell = require('shelljs');
-var ncp = require('child_process');
-
-//------------------------------------------------------------------------------
-// shell functions
-//------------------------------------------------------------------------------
-var shellAssert = function () {
-    var errMsg = shell.error();
-    if (errMsg) {
-        throw new Error(errMsg);
-    }
-}
-
-var test = function (options, p) {
-    var result = shell.test(options, p);
-    shellAssert();
-    return result;
-}
-exports.test = test;
-
-var mkdir = function (options, target) {
-    if (target) {
-        shell.mkdir(options, target);
-    }
-    else {
-        shell.mkdir(options);
-    }
-
-    shellAssert();
-}
-exports.mkdir = mkdir;
-
-var assert = function (value, name) {
-    if (!value) {
-        throw new Error('"' + name + '" cannot be null or empty.');
-    }
-}
-exports.assert = assert;
-
-var cp = function (options, source, dest) {
-    if (dest) {
-        shell.cp(options, source, dest);
-    }
-    else {
-        shell.cp(options, source);
-    }
-
-    shellAssert();
-}
-exports.cp = cp;
-
-var run = function (cl, options, noHeader) {
-    if (!noHeader) {
-        console.log();
-        console.log('> ' + cl);
-    }
-
-    var rc = 0;
-    var output;
-    try {
-        output = ncp.execSync(cl, options);
-    }
-    catch (err) {
-        if (!options.inheritStreams) {
-            console.error(err.output ? err.output.toString() : err.message);
-        }
-
-        process.exit(1);
-    }
-
-    return (output || '').toString().trim();
-}
-exports.run = run;
-
-var matchFind = function (pattern, root, options) {
-    assert(pattern, 'pattern');
-    assert(root, 'root');
-
-    // create a copy of the options
-    var clone = {};
-    Object.keys(options || {}).forEach(function (key) {
-        clone[key] = options[key];
-    });
-    options = clone;
-
-    // determine whether to recurse
-    var noRecurse = options.hasOwnProperty('noRecurse') && options.noRecurse;
-    delete options.noRecurse;
-
-    // normalize first, so we can substring later
-    root = path.resolve(root);
-
-    // determine the list of items
-    var items;
-    if (noRecurse) {
-        items = fs.readdirSync(root)
-            .map(function (name) {
-                return path.join(root, name);
-            });
-    }
-    else {
-        items = find(root)
-            .filter(function (item) { // filter out the root folder
-                return path.normalize(item) != root;
-            });
-    }
-
-    return minimatch.match(items, pattern, options);
-}
-exports.matchFind = matchFind;
-
-var banner = function (message, noBracket) {
-    console.log();
-    if (!noBracket) {
-        console.log('------------------------------------------------------------');
-    }
-    console.log(message);
-    if (!noBracket) {
-        console.log('------------------------------------------------------------');
-    }
-}
-exports.banner = banner;
-
-var fail = function (message) {
-    console.error('ERROR: ' + message);
-    process.exit(1);
-}
-exports.fail = fail;
-
-var ensureExists = function (checkPath) {
-    assert(checkPath, 'checkPath');
-    var exists = test('-d', checkPath) || test('-f', checkPath);
-
-    if (!exists) {
-        fail(checkPath + ' does not exist');
-    }
-}
-exports.ensureExists = ensureExists;
-
-// Validates the structure of a task.json file.
-var validateTask = function (task) {
-    if (!task.id || !check.isUUID(task.id)) {
-        fail('id is a required guid');
-    };
-
-    if (!task.name || !check.isAlphanumeric(task.name)) {
-        fail('name is a required alphanumeric string');
-    }
-
-    if (!task.friendlyName || !check.isLength(task.friendlyName, 1, 40)) {
-        fail('friendlyName is a required string <= 40 chars');
-    }
-
-    if (!task.instanceNameFormat) {
-        fail('instanceNameFormat is required');
-    }
-};
-exports.validateTask = validateTask;
