--- Extensions/XplatGenerateReleaseNotes/package.json
@@ -1,24 +1,0 @@
-{
-    "name": "XPlatGenerateReleaseNotesTask",
-    "scripts": {
-        "installTasks": "node make.js installTasks",
-        "build": "node make.js build",
-        "clean": "node make.js clean",
-        "package": "node make.js package"
-    },
-    "devDependencies": {
-        "minimist": "^1.2.6",
-        "ncp": "^2.0.0",
-        "path": "^0.12.7",
-        "rimraf": "^2.6.2",
-        "semver": "^5.5.0",
-        "shelljs": "^0.8.5",
-        "tslint": "^5.20.1",
-        "typescript": "^3.5.3",
-        "validator": "^13.7.0"
-    },
-    "dependencies": {
-        "child_process": "^1.0.2",
-        "handlebars": "^4.7.7"
-    }
-}
