+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/build-consumed-artifact-table.md
