+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/build-handlebars-template.md
