+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/count_workitems_by_type.js
