+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/date_formatter.js
