+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/get_only_message_firstline.js
