+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/get_unique_projects.js
