+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/gitflow-extension.js
