+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/lookup_PR_by_LastCommitId.js
