+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/release-handlebars-dump-template.md
