+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/release-handlebars-template.md
