+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/replace_text.js
