+++ SampleTemplates/XplatGenerateReleaseNotes (Node based)/Version 3 (and later)/return_parents_only.js
